requirements.txt
paho-mqtt
requests
streamlit==1.36.0
pandas==2.2.2
supabase
joblib==1.4.2
scikit-learn==1.5.0

